Implement upload split

+++ RMS/ArchiveDetections.py
@@ -257,49 +257,61 @@ def archiveDetections(captured_path, archived_path, ff_detected, config, extra_f
 
     if file_list:
 
+        # Initialise archive names
+        archive_name, imgdata_archive_name, metadata_archive_name = None, None, None
+
         # Create the archive ZIP in the parent directory of the archive directory
         archive_base = os.path.join(os.path.abspath(os.path.join(archived_path, os.pardir)),
             os.path.basename(captured_path))
 
-        archive_name = f"{archive_base}_detected"
 
-        # Archive the files
-        log.info(f"Generating archive file {archive_name}")
-
-        archive_name = archiveDir(captured_path, file_list, archived_path, archive_name, \
-            extra_files=extra_files)
+        # These sets will only be used in the upload_split code path, but keeps the conditional
+        # sections shorter
 
         # Create the imgdata set which is the union of the sets of FF files and FR files
         imgdata_set = (set([item for item in file_list if item.startswith("FF") and item.endswith(".fits")]) |
                        set([item for item in file_list if item.startswith("FR") and item.endswith(".bin")]))
 
-        # Create a directory to hold the imgdata files, archive, then remove the directory
-        imgdata_archived_path = f"{archived_path}_imgdata"
-        imgdata_archive_name = f"{archive_base}_imgdata"
-        log.info(f"Generating archive file {imgdata_archive_name}")
-        imgdata_archive_name = archiveDir(captured_path, imgdata_set, imgdata_archived_path,
-                                          imgdata_archive_name, extra_files=extra_files)
-
-        if os.path.exists(imgdata_archived_path):
-            if os.path.isdir(imgdata_archived_path):
-                shutil.rmtree(imgdata_archived_path)
-
         # Create the metadata set which is the relative complement of file_list in imgdata_set
         # or the set of files in file_list which are not in imgdata_set
         metadata_set = set([item for item in file_list if item not in imgdata_set])
 
-        # Create a directory to hold the metadata files, archive, then remove the directory
+        # Create all the required archive names and paths
+        archive_name = f"{archive_base}_detected"
         metadata_archived_path = f"{archived_path}_metadata"
         metadata_archive_name = f"{archive_base}_metadata"
-        log.info(f"Generating archive file {metadata_archive_name}")
-        metadata_archive_name = archiveDir(captured_path, metadata_set, metadata_archived_path,
-                                           metadata_archive_name, extra_files=extra_files)
+        imgdata_archived_path = f"{archived_path}_imgdata"
+        imgdata_archive_name = f"{archive_base}_imgdata"
 
-        if os.path.exists(metadata_archived_path):
-            if os.path.isdir(metadata_archived_path):
-                shutil.rmtree(metadata_archived_path)
+        if not config.upload_split:
 
+            # Archive the files
+            log.info(f"Generating archive file {archive_name}")
 
+            archive_name = archiveDir(captured_path, file_list, archived_path, archive_name, \
+                extra_files=extra_files)
+
+        else:
+
+            # Do the metadata first so that might be uploaded first, reducing observation -> processing time
+
+            # Create a directory to hold the metadata files, archive, then remove the directory
+            log.info(f"Generating archive file {metadata_archive_name}")
+            metadata_archive_name = archiveDir(captured_path, metadata_set, metadata_archived_path,
+                                               metadata_archive_name, extra_files=extra_files)
+
+            if os.path.exists(metadata_archived_path):
+                if os.path.isdir(metadata_archived_path):
+                    shutil.rmtree(metadata_archived_path)
+
+            # Create a directory to hold the imgdata files, archive, then remove the directory
+            log.info(f"Generating archive file {imgdata_archive_name}")
+            imgdata_archive_name = archiveDir(captured_path, imgdata_set, imgdata_archived_path,
+                                              imgdata_archive_name, extra_files=extra_files)
+
+            if os.path.exists(imgdata_archived_path):
+                if os.path.isdir(imgdata_archived_path):
+                    shutil.rmtree(imgdata_archived_path)
 
         return archive_name, imgdata_archive_name, metadata_archive_name
